recsys/requirements.txt
catboost==1.2
faiss-cpu==1.7.4
gensim==4.3.2
huggingface-hub==0.17.3
implicit==0.7.0
keras==2.13.1
lightfm==1.17
matplotlib==3.7.3
nltk==3.8.1
numpy==1.24.3
optuna==3.3.0
pandas==2.0.3
plotly==5.17.0
polars==0.18.6
pyarrow==12.0.1
redis==4.6.0
scikit-learn==1.3.0
scikit-surprise==1.1.3
scipy==1.10.1
torch==2.1.0
torch_geometric==2.5.2
tqdm==4.65.0
